Split scream_std_meta.hpp into multiple header, and added enable_shared_from_this
Splitting makes maintenance of the code easier: less code to look at in every
single file, and if/when a feature becomes available in std, simply remove
that single file.

util::enable_shared_from_this mimics the c++17 implementation of the std version.
Note that std::enable_shared_from_this is already present in c++11, but lacks
an important feature, as described in the comments in the file.