Set default NTEST_PARALLEL_JOBS=MAX_MPITASKS_PER_NODE and limit E3SM machines
This resets the default value of NTEST_PARALLEL_JOBS to MAX_MPITASKS_PER_NODE
so as to not make any behavioral changes to CESM.

Warning: This is not a safe value on machine with batch systems who's
login nodes are more limited than the compute nodes and therefore
NTEST_PARALLEL_JOBS should be set on these systems.

@jgfouca found via E3SM testing that limiting to 4 parallel jobs was
required for many of the testing machines with batch systems to prevent
hammering login nodes. Therefore, we set that value for these E3SM
machines:
* cori-haswell
* cori-knl
* blues
* anvil
* bebop
* theta
* titan
* summit

Warning: Non test machines for E3SM that have a batch system may still
oversubscribe parallel test jobs.