Use truncated division operator in SEQ test for py2+3
When using python 3, the SEQ test will fail to set NTASKS_* because
the XML set_value expects an integer, but a float is produced. This is
because of the change from the `/` operator using truncated division in
python 2 to using real division in python 3. Simply replacing `/` with
the truncated division operator `//` fixes the issue (available since
python 2.2).