Changed how cubed_sphere_map is set
I erroneously made cubed_sphere_map show up in the namelist, but the correct approach was to modify it in the code and hide this from the user. It is now set to 2 in namelist_mod.F90 when fv_nphys>0. I also added formula for setting ncol in horiz_grid.

	modified:   cime/config/e3sm/config_grids.xml
	modified:   components/cam/bld/build-namelist
	modified:   components/cam/bld/config_files/horiz_grid.xml
	modified:   components/cam/bld/namelist_files/namelist_defaults_cam.xml
	modified:   components/homme/src/share/namelist_mod.F90