Homme: fixed link error and generalized ColumnOps
The link error is the usual undefined reference for static constexpr
class member that are odr-used. It is fixed in C++17, but, as usual,
we can't rely on anything more recent than C++11.

In ColumnOps we templated all methods on the input type. It can be anything
that implements operator()(int)->Scalar. Hence, 1d views and functors are
both fine.