COMPOSE: Fix bug in nearest-point calculation when halo > 1.
If semi_lagrange_nearest_point_lev > 0, and with the default halo of 2, there
was a bug in the nearest-point calculation. The nearest point was being computed
w.r.t. a 1-halo perimeter, but it should be computed w.r.t. the actual 2-halo
perimeter. This was a result of initializing the nearest-point data before
extending the halo region. This commit fixes this bug.

The bug in practice is benign because the 2-halo is intended to make the
nearest-point calculation unnecessary, and in practice it does. But we want it
to be correct, of course, in case of a very large qsplit setting. The purpose of
the nearest-point calculation is to permit reasonable qsplit to be set to an
optimistically large value and have advection proceed with reasonable accuracy
rather than crash on out-of-halo departure points, when the user wants to push
SL advection hard.