Change default value for config_blend_bdy_terrain to false
Regional MPAS-A simulations using real-data initial conditions will likely
involve two stages to the generation of initial conditions: (1) interpolating
static fields, and (2) interpolating atmospheric initial conditions.

The blending of boundary terrain is expected to be combined with the second
stage (since the terrain to blend with should match the terrain in the dataset
used for ICs), so it makes more sense to have the default for
config_blend_bdy_terrain as false, so that users will only need to set this
option to true in the second stage of generating ICs.

Previously, with the default set to true, users would likely need to first
change the value to false before interpolating static fields, then change
the value back to true before interpolating atmospheric ICs.