This is an archive tag for reduced precision paper submitted to JAMES.
The model is used for to test the impact of reduced precision on E3SM atmosphere model(EAM)'s simulation. In this model configuration, A "round down and up" procedure where the lower order mantissa of a double-precision number was truncated for all the prognostic variables used by the turbulence and clouds parameterization CLUBB. Thus, half-precision (with 10 significand bits) was used for the turbulence and clouds parameterization CLUBB while double precision was used for the rest of the model.

The simulations include:
1. The convergence test simulations (1h - 5day)
2. 10-year climate simulations for E3SM atmosphere model (EAM).