Move domain_ptr and corelist from module variables to arguments in subdriver
Towards the goal of supporting multiple instances of domains within a single
simulation, this commit moves domain_ptr and corelist from module variables
in the mpas_subdriver module to arguments to the mpas_init, mpas_run, and
mpas_finalize routines.

The main program (in mpas.F) now passes instances of a core_type and
a domain_type to the subriver's mpas_init, mpas_run, and mpas_finalize routines.