Add a stand_alone test for P3 based on SCREAM-AD infrastructure.
This commit replicates the "ping-pong" test to construct a stand
alone test of P3 using the SCREAM-AD infrastructure.

Some important notes at this point:
The test itself doesn't use the P3Microphysics process constructed in
the atmosphere_microphysics.cpp code.  Instead it has a local DummyProcess
just like the ping-pong test, but this one uses the scream_p3_interface.hpp
code and the scream_p3_interface.F90.

This wsa done because there was an error when using the atmosphere_microphysics.cpp
code as it is currently coded up that I could not figure out.  The error
allowed the code to compile, but got a segmentation fault during the
factory build phase of the stand alone test.

A future commit will fix the atmosphere_microphysics.cpp code and change the
stand alone test to use this code instead of a local "DummyProcess"