Update gravity wave drag mods to ingest improved tuning for QBO via nml
Convective heating rate conversion factor (CF) and gravity wave tendency efficiency
for Beres scheme (effgw_beres) are retuned to improve QBO for adoption in EAMv2.

In the new tuning with much improved QBO, the conversion factor is
changed to 12.5 from 20.0, and the tendency efficiency is reduced to 0.35 from 0.4.

This implementation adds namelist control for the conversion factor to incorporate the
new tuning for V2 while preserving reproducibility of all existing compsets.
The new values for CF and effgw_beres will be ingested via new compsets and corresponding use case files.

[BFB]