Changed cpp options to avoid incorrect fortran output.
cpp is used by the chemistry pre-processor to manipulate the fortran
output. It seems that modern versions of cpp behave differently, so the
chemistry pre-processor has problems and produces different fortran
output. There are two main issues:

1) cpp adds its copyright information to the top of the output using
   C-type comments, which confuses the chemistry pre-processor (and
   probably the E3SM Fortran compiler).

2) cpp deletes parts of the fortran output because it misidentifies it
   as a comment.

Issue 1 occurs with the -C option to cpp, and issue 2 occurs without
the -C option to cpp.  This commit solves the problem by replacing the
-C option with -traditional-cpp.

[E3SM is BFB.  Output of chemistry pre-processor is changed]