Making units specification part of the FieldIdentifier
Since Units does not have a default constructor, one is "forced" to provide
units during the construction of the identifier. I could have used some
sort of "invalid" units (like [0] non-dimensional), but I thoguht specifying
the units of a field was as important as specifying its name, so it could
not be deferred to a post-construction set method.

Notes:
 - inside all the unit tests, all the units are picked randomly.
 - in the atmosphere dynamics process (the interface to homme), I set the
   units of states and forcings. In particular, the units for the forcing
   are those used by ftype 0,2 and 4 in homme. If ftype is 1, they would
   be wrong. I put a check to make sure the ftype is 0, 2 or 4 for now.