Upgraded ping pong test to use two different grids
The two grids are dummy: the second one is equal to the first backwards.
Nevertheless, it triggers AtmosphereProcessGroup to build a remapper
(which is also dummy, an ad-hoc implementation for the two mirror grids).

While upgrading the test, I fixed a bug in the time stamp print method,
as well as reformatted its output. I also had to introduce small changes
in the abstract remapper and fix the atm process group.