Specify recursive doubling for MPI_Allreduce on Compy
The default algorithm used for MPI_Allreduce with available
versions of the Intel MPI library is failing for some process
counts and problem sizes. This default appears to be a hybrid
algorithm that uses Rabenseifners's algorithm when the number
of processes becomes large, and standalone experiments indicate
that the implementation of Rabenseifner's algorithm is at fault
here. A workaround while Intel addresses this error is to
specify a different algorithm using an environment variable
set in env_mach_specific.xml for Compy when using the Intel
MPI library.