Back out all changes except for scatter_global_dbl
In E3SM runs, CICE is used only in the F case, in which case
the only routine modified in the previous commit is
scatter_global_dbl. As this routine is the only one that needs
to be modified in order to address the larger message tag failures
in E3SM, and as this is the only routine that we can check when testing
using E3SM, the changes to the other routines in ice_gather_scatter.F90
are backed out.

BFB