Adds explicit types to literals in COARE ocean surface flux scheme.
The COARE ocean surface flux scheme had a large number of un-typed
literals (including some in functions that caused a build issue on
Summit). This change converts most to "_R8". This shouldn't affect
the climate of "out of the box" cases (which use a different ocean
surface flux scheme), but may be non-BFB with the original COARE
implementation. This needs to be tested.

[BFB] - Bit-For-Bit
? [FCC] - Flag Climate Changing

Issue number 3105