Bug fix in phys-dyn remapper
The fwd remap (phys->dyn) requires a halo exchange (export unique cols to
the repeated partition of SE elems structure). During the local remap, not
all the entries in the dyn view are overwritten, so they would have some
leftover garbage from before. To avoid this, we need to explicitly zero-out
the view before the remap.