Use hy(a,b)m and ps_v to compute p rather than dp3d.
check_p reports
    atm.log: gfr> ERROR dp p  0.000000000000000E+000  2.651938505852153E-011
in an ne30pg2 FC5AV1C-L run, meaning dp agrees exactly with dp3d but
these two defs of p agree only to ~10 digits:
    do k = 1,nlev
       p(:,:,k) = hvcoord%hyam(k)*hvcoord%ps0 + hvcoord%hybm(k)*ps
    end do
and
    dp_accum(:,:,1) = hvcoord%hyai(1)*hvcoord%ps0
    do k = 1,nlev
       dp_accum(:,:,2) = dp_accum(:,:,1) + dp3d(:,:,k)
       p(:,:,k) = half*(dp_accum(:,:,1) + dp_accum(:,:,2))
       dp_accum(:,:,1) = dp_accum(:,:,2)
    end do
This is because hybm or hyam do not give exactly the middle of the
element. The first def is consistent with what the EAM driver
currently does, so I'm switching back to it.