Apply temperature clipping for RRTMGP
aquaplanet runs tend to exhibit temperatures that are too low for RRTMGP in the
upper levels, so slipping was added when aqua_planet is true. This is a
temporary fix to facilitate the development of the RCE compset.

A new RCE ozone input file was added to the RCE use case.

	modified:   components/cam/bld/namelist_files/use_cases/RCEMIP_EAMv1.xml
	modified:   components/cam/src/physics/rrtmgp/radiation.F90