Table lookup bugfix for P3
This PR fixes an issue where the hard-coded p3 table version was set to version 2.8.2, which doesn't reflect the recent change to v4 of the table.

Additionally there was a bug associated with the new rtype8 variable from p3_utils that wasn't defined if the build type was not SCREAM. This broke the standard E3SM build.

TODO comments have been added addressing the desire to eventually make the lookup table a runtime option instead of hard-coded.