Add do_aerosol_rad option and move constant_albedo
The constant_albedo was originally put directly into RRTMGP, but I
realized that was dumb, so I moved it to atm_import() so I could
directly modify the albedo values as they come out of the coupler. This
way they will be constant for all components of the ATM model.

	modified:   components/cam/bld/build-namelist
	modified:   components/cam/bld/namelist_files/namelist_defaults_cam.xml
	modified:   components/cam/bld/namelist_files/namelist_definition.xml
	modified:   components/cam/bld/namelist_files/use_cases/RCEMIP_EAMv1.xml
	modified:   components/cam/src/control/physconst.F90
	modified:   components/cam/src/cpl/atm_import_export.F90
	modified:   components/cam/src/physics/rrtmgp/radiation.F90