Homme: Removed the DiagnosticsBase class.
The code duplication saving was little, and the theta model actually started
to use more diagnostics times, which made the construction of DiagnosticsBase
more complicated. I realized it was just easier to duplicate code in
prim_diag_scalars and be done with it.

Incidentally, Diagnostics was the only class that used/needed the method
'create_with_base' in the Context class. So I purged that method.