Update clubb to print tunable parameters and make TTEND consistent
clubb_intr.F90 is updated to record the CLUBB tunable parameter values
in the atm.log file.

Also change is to replace STEND_CLUBB with TTEND_CLUBB. Previously
the quantity in STEND_CLUBB is inconsistent with the name and the unit
attributes. This is now corrected using  new variable TTEND_CLUBB.

[BFB]