Homme, nhxx: Make zeroulpn'ed values fully sensitive.
Luca pointed out that an effectively reduced-precision z=pow(x,e) could mask BFB
diffs in x. So add a third arg to zeroulpn, 'replace', that in this case is
replace=x. The n least significant bits of x's fraction part are inserted into
those of the zeroulpn output. A bfb_pow built on this makes z sensitive to all
its of x.