Homme: correctly detecting if nvcc is used
The previous way assumed that nvcc_wrapper was the CXX compiler.
In reality, most likely CMAKE_CXX_COMPILER=mpicxx.
However, we SHOULD find that nvcc_wrapper is the backend
C++ compiler for mpicxx (which, with OpenMPI can be custom
set via the OMPI_CXX env variable). To detect this, one
can simply run ' --version'. If the
backend compiler is indeed nvcc_wrapper, the output WILL
contain the string 'nvcc' somewhere.