Homme: added XX_BFB_TESTING guards
Also, some uses of XX_NONBFB_COMING were inappropriate,
and XX_BFB_TESTING was the correct call.

In particular, when xx choses a different arithmetic path
than the f90 code (for performance/safety reasons), only
one of the two should keep both path (and activate the
"non native one" only in bfb tests).

The other case where XX_BFB_TESTING is needed is when using
math functions, whose impl is not necessarily bfb on GPU.
For now, this is manily just std::pow, which, in bfb builds,
needs to be specially handled, to ensures bfb (with *very*
high probability) of pow between CPU and GPU.