Homme: slight change to infrastructure_ut
The bfb_pow function from BfbUtils was not really tested.
Instead, we had a verbatim copy, with the exception that
the zeroulpn was always called. However, due to CPP macros
errors, the one in BfbUtils was not behaving in the right
way, and the unit test was not exposing it (cause it wasn't
testing that routine). Now infrastructure_ut calls bfb_pow
from BfbUtils.hpp.