Fix a bug in P3 microphysics which still relied on the current nstep (var=it) to determine if an action would be done.
This logic was originally in P3 because P3 handled the saturation
adjustment and had to take different actions depending on whether
or not it was the first step (it=1) or any other step (it>1).

In scream we removed the saturation adjustment from p3 so this is
no longer a distinction that is needed.

Should be BFB, but presumably could see a slight change to the first
step.  And of course any subsequent step because of the propagation
of that difference.