Change the parameters for p3_run_and_cmp test.
The original set of parameters in the p3_run_and_cmp test exercised the
'it' variable.  Changes to P3 have made the solution no longer dependent
on this variable.

This commit uses the 'it' variable differently, specifying a number of
iterations of p3_main.  This commit also switches the timestep to the
standard P3 timestep of 300s.

For the cases where there is more than one iteration the write or read
command is issued for every iteration providing a BFB test for multiple
time slices.