Setup interface between MCT and SCREAM to run as a "DEAD" model for ATM.
This step in building SCREAM inside CIME replaces the old atm_comp_mct
interface with a slightly more sophisticated "DEAD" implementation of
the ATM model.

With this change it is possible to now run a simulation, albeit one where
nothing happens.  The next step will be to have the interface actually run
SCREAM code within MCT.