Homme: fixing accidental lucky initialization of buffers in BFB testing
Catch aboud malloc: it does not init the memory, but if the memory comes
from the OS, the OS will zero it out before handling it over (assuming
security protections are enabled in the OS). This can cause tests to pass
even if buffers were not correctly init-ed.

E.g., if the functor sums into a buffer without zeroing out (and we test
it only once). More concretely, Dirk was checking that an entry was indeed
zero. It was in the unit tests, but not in actual runs (since the buffers
were first used by other functors, which polluted the values).