Homme: fixing bfb end-end runs in f90
When running with HOMMEXX_BFB_TESTING on, the f90 target relies
on the kokkos impl of the tridiag solver, to get bfb results.
However, this requires f90 to initialize/finalize kokkos.
The best place to do this is prim_main. However, one needs
to be careful, since the kokkos targets will already try
to init kokkos. To solve this impasse, I added a KOKKOS_TARGET
macro in the config.h.cmake.in files, which is 0 for f90
targets, and 1 for kokkos targets.