Fix p3_stand_alone and p3_shoc_coupled unit tests.
This commit creates a new atmospheric process for initialization of
microphysics variables, SAMicrophysics, referring to
"stand-alone-microphysics".  It is a temporary fix to handle an issue
that arises from running p3 stand alone which requires variables declared
as "required" or "input" to be somehow initialized without calling other
processes.

The fix is to include a stand-alone process which treats all variables as
computed, thus allowing for the changes to the variable to be accessible to
the main p3 microphysics call.  The stand-alone process only has an init call
defined since it isn't expected to do anything during the run and finalize
sequence.

This commit also removes an old compiler flag which did not build the p3
and p3_shoc stand-alone tests if SCREAM_FPE was on.

[BFB]*

*Although there are now two new twos in the debug_nopack_fpe suite owing
to p3 and p3_shoc tests now being turned on.