Add critical sections around calls to shr_sys_irtc
The CIME utility routine shr_sys_irtc is not thread safe. This causes
errors in the logic calculating the cost per chunk when threading
over the chunk loops in physpkg.F90 (six locations), confirmed on
both Cori-KNL and Compy. This error is only obvious when the
user_nl_cam option

 phys_chnk_cost_write = .true.

is set, but it is occurring in every threaded run (just not causing
any other problems, that we are aware of).

To make shr_sys_irtc thread safe requires putting the entire body of
the function into a critical section, including the call to
system_clock (at least on Compy). Note that even if the call to
system_clock was thread safe, there are also local saved variables
that would need to be protected in a critical section.

Rather than trying to make shr_sys_irtc thread safe, here we wrap the
three calls to shr_sys_irtc in physpkg in named critical sections. If
it is decided to modify shr_sys_irtc in the future, there critical
sections can be removed then.

[BFB]