Add atmospheric physics decomposition diagnostics
In recent investigations into the performance impact on atmospheric
physics of changing the PCOLS setting, it was determined that output
of some physics decomposition specifics was very useful, i.e., the
total number of columns in the atmospheric physics, the total number
of chunks, the minimum and maximum number of physics columns assigned
to each chunk, and the minimum and maximum number of physics columns
assigned to each process. As optimizing over PCOLS will likely become
more common in the future, this output is added at the end of
phys_grid_init. The format of the immediately preceding phys_grid_init
output is also modified, to be consistent with the new output (and to
be easier to read).

[BFB]