Add threadprivate declaration and remove critical sections
The function shr_sys_irtc can be made thread safe in the current usage
context by adding

!$omp threadprivate(last_count,count_offset)

where last_count and count_offset are the two local saved variables.

With this change, the critical sections around the calls to
shr_sys_irtc in physpkg.F90 are no longer necessary, and are removed
here.

[BFB]