Move temperature checks after setting rad state
Move temperature checks to after we set the input rad state, rather than
using the input physics_state object. This is because we ultimately want
to provide the option of clipping temperatures to keep them in bounds if
we need to, but we cannot modify the physics state object directly.