Scream: added --keep-tree flag
The flag allows to test an unclean repo against pre-built baslines
(stored in the directory passed with --baseline-dir) or against
the current branch itself (that is, the user specified -b HEAD).
If the user specified a different baseline ref, then using the
--keep-tree flag is invalid, and an error is raised.

Notice, that if the user has a non-clean repo and wants to test
against HEAD, then the user must first commit/stash the changes,
just like when testing against any other branch.