Move state/pbuf out of cloud optics routines
Finish moving state and pbuf all the way out of cloud optics routines.
This moves handling of state and pbuf all the way up to the shortwave
and longwave drivers. This is a significant change in how the optics
routines are called though; in the shortwave and longwave drivers, we
now call not just the single `set_cloud_optics_` routines to handle all
of the getting of band optics, McICA sampling, and daytime compression,
but instead now call separate routines to handle each of these tasks.
This is an improvement though, because it further breaks up the
radiation calls into separate concerns that are more compact and easier
to follow (and easier to test). Note that pbuf and state *are* still
used in the aerosol optics portions of the `cam_optics` modules. We also
still have the `do_snow_optics` function that depends on pbuf, which we
will want to move up to a higher level.