Add ability to parse index files of static data sets
This commit creates a new module, mpas_parse_geoindex, which can parse index
files that describe static data sets. The index files and data sets are the
same data sets that are used by WRF and the WPS. For more information on valid
fields for an index file please see section 3-53 "Description of Index Options"
of the WRF Users Guide.

The mpas_parse_geoindex is currently not used by any portion of the
init_atmosphere core; it is only compiled.