Add support for copy_file tag
This is needed for copying config files that may then be altered
by the user within a test case.

Since there is a significant amount of redundant code between this
and and support for the `add_link` tag, a helper function has been
added to encapsulate the redundancy.

Some error messages have been modified to raise exceptions.