Eliminate compile-time dependence on pcols in camsrfexch.F90
The data types cam_in_t and cam_out_t in the camsrfexch module
in control/camsrfexch.F90 refer to pcols in dimensioning some of
the fields in these types. These fields are changed to allocatable.
Then these fields for variables of type cam_in_t are allocated
in hub2atm_alloc and deallocated in hub2atm_dealloc, and these
fields for variables of type cam_out_t are allocated in atm2hub_alloc
and deallocated in atm2hub_dealloc. All of these routines reside
in control/camsrfexch.F90.

This removes one dependency on the compile-time specification
of pcols.

[BFB]