Eliminate compile-time dependence on pcols in check_tracers logic
The data type check_tracers_data in the check_energy module, in
physics/cam/check_energy.F90, refers to pcols in dimensioning some
of the fields in the type. These fields are changed to allocatable,
and these fields for variables of type check_tracers_data are then
allocated in check_tracers_init, and deallocated in a new subroutine
check_tracers_fini. A call to check_tracers_fini is added to
subroutine tphysac and to subroutine tphysbc, both in
physics/cam/physpkg.F90.

This removes one dependency on the compile-time specification
of pcols.

[BFB]