Eliminate compile-time dependence on pcols in zm_conv
The module variable dcapemx in the zm_conv module, in
physics/cam/zm_conv.F90, is dimensioned as size pcols.
This is changed to allocatable, and is allocated in
subroutine zm_convr. It is not subsequently deallocated as
dcapemx is persistent throughout the execution of EAM.
A test in zm_convr prevents it from being allocated more
than once.

This removes one dependency on the compile-time specification
of pcols.

[BFB]