Eliminate compile-time dependence on pcols in diag_surf
The line

    call outfld('QFLX',     cam_in%cflx(1,1), pcols, lchnk)

in subroutine diag_surf in physics/cam/cam_diagnostics.F90
fails to compile when pcols is not a parameter. However,

    call outfld('QFLX',     cam_in%cflx,      pcols, lchnk)

compiles successfully when pcols is a global runtime variable,
and appears to still do the right thing. This will be verified
in future, more extensive, testing.

This removes one dependency on the compile-time specification
of pcols.

[BFB]