Implement pcols as a runtime option in EAM
In the earlier 2000's, empirical experiments indicated that setting
the first dimension of the chunk data structure at compile-time
allowed the compiler to generate more performance-efficient code, and
this motivated this design choice in CAM. However, CAM/EAM has evolved
significantly since then, and, for example, physics_types now uses
psetcols to define this first dimension, and the compiler never sees a
compile-time parameter. So compile-time setting of PCOLS likely no
longer serves a useful purpose.

Here pcols is implemented as a runtime option, currently specified
via the user_nl_cam namelist variable phys_chunk_fdim. phys_chunk_fdim
must be at least as large as 1, but there is no specified upper
bound. The current compile-time value of PCOLS is used to define the
default value for phys_chunk_fdim (and pcols). PCOLS is now
represented in EAM as the parameter ppcols, but is only referenced
in defining the default for phys_chunk_fdim.

[BFB]