SCREAM: minor feature upgrade to test-all-scream
We currently only do integration testing against a baseline ref.
If/when the set of tests gets larger, we may want to test against
a set of pre-build baselines, to keep the testing faster.
In such a scenario, we still need to merge origin/master in the
tested branch, to ensure correctness of master if/when the pr
branch gets merged.