Implementation of 1990(20TR) defaults for superfast chemistry.
Implement the defaults for a 1990 simulation. It is setup as a
transient simulation starting in 1990.  It mostly uses time-series
input files, but there are a couple of files for which a 1990-era file
is used because it was all that was easily available.

The inputs are from CMIP6 where possible, and otherwise from the AR5
era.

[BFB, except for superfast 20TR compsets]