Automatically rebuild api documentation whenever 'make html' is run
This way we don't need to remember to rerun 'make api' before running
'make html'. This is also needed when using
https://github.com/ESMCI/versioned-doc-builder with the '--clean' flag
(with the last commit, that made 'make clean' also run 'clean_api'):
otherwise, the api documentation gets cleaned and never rebuilt in time
to make the html documentation.

This makes the new 'clean_api' somewhat less important, but it still
seems good to have that in place, so that a 'make clean' also cleans out
the generated api documentation.