Bug fix for COSP2 MODIS simulator
This is to incorporate bug fix by Dustin Swales for MODIS simulator
in COSP2 as described at https://github.com/ESCOMP/CAM/pull/95.

Issue 1:
From COSP1.4 to COSP2 there was a change in the LUTs used by the particle-size
retrieval within the MODIS simulator, but this change was not accounted for on
the EAM side. This mismatch in LUTs resulted in the retrieval failing more often,
hence the reduction in MODIS total cloud fraction.
Solution 1:
Use the new LUTs in the COSP/EAM interface (cosp2/optics/cosp_optics.F90).

Additionally there was a bug-fix to the MODIS optics in COSP2 that hadn't
yet been propagated into EAM. The implementation of COSP2 in EAM occurred
around the same time when this fix was implemented in COSP2
(and as a patch to COSP1.4)

Issue 2: The combining of the MODIS optical fields, single-scattering
albedo and asymmetry parameter, were incorrect.
Solution 2: Propagate changes in the computation of optical properties into CAM.

[BFB] except for COSP-MODIS diagnostics