Add option to restore compile-time setting of pcols
Only a small change is required to restore the ability to set
pcols at compile-time. This allows the performance comparison
between compile-time and runtime pcols while using the new code
required for runtime pcols. Experiments indicate that the
performance differences in ATM_RUN are primarily due to
compile-time vs. runtime setting of pcols, and not due to
this (small amount of) new code.

To enable compile-time setting of pcols, add '-DPPCOLS' to
CPPFLAGS in Macros.cmake for a case before building. With this
change, phys_chunk_fdim will be ignored (but a warning message
will be output if phys_chunk_fdim differs from the compile-time
setting of pcols.

[BFB]