Change member in cosp_optical_inputs type to be allocatable
The COSP2 cosp_optical_inputs type includes a member that is of type
radar_cfg. radar_cfg itself contains members that are large, which causes
performance issues when allocating a variable of type
cosp_optical_inputs (implicitly as a local variable or explicitly with
an allocate). This can be resolved by instead defining the member
of type radar_cfg in type cosp_optical_inputs to be allocatable and
explicitly allocating/deallocating this member when using variables of
type cosp_optical_inputs.

[BFB]