Change in config_compilers.xml
When handling netcdf paths, we should consider C and Fortran
separately. This is more flexible, and can accommodate also
the case where they are both in the same path (simply define
NETCDF_C_PATH and NETCDF_FORTRAN_PATH to be the same),
while a single NETCDF_PATH env var makes it hard for those
who have two separate folders for the C and Fortan installations.