Some cleanup following maint-5.6 merge
maint-5.6 added support for new OMIP compsets in POP as well as better handling
of the different versions of JRA forcings. Unfortunately, a lot of these
changes conflicted with other updates to master and the original merge couldn't
be tested because the POP changes weren't available on master yet. I brought
the POP changes from cesm2_1_x_rel to master, and these were the CIME changes
necessary to run the aux_pop testlist:

1. Support the JRA025v2 grid
2. Properly configure POP for compsets with longnames that start with OMIP
3. Support DATM_CO2_TSERIES=omip
   -- required converting fco2_datm_global_ssp585_simyr_1750-2020_CMIP6 file
      from netcdf-4 to cdf5
4. Remove CORE_IAF_JRA.CORE2.ArcFactor
5. Remove extra rof.iaf_jra block from namelist_definition_drof.xml
6. Update CCSM_CO2_PPMV default value for POP ecosystem from 284.7 to 284.317
   unless specificing ECOCESM20