Bring back the setting of dlnd strm_domdir and strm_domfil to null
When dlnd was changed to support a new mode for MOSART in
ESMCI/cime#3457, the default settings of null for strm_domdir and
strm_domfil were accidentally removed. It turns out that these explicit
settings to null are important for the logic in nmlgen.py:
create_stream_file_and_update_shr_strdata_nml.

This commit brings back these settings to null