Disallow inflow to hydro system from outside the glacier
This commits eliminates inflow from ocean to glacier or land to glacier.
Outflow is still allowed.
For ocean this is invalid because ocean water has a different density.
For land this would only happen if there is a supply of water, which is not currently handled.
Do this by simply zeroing the hydropotential gradient in those cases before calculation of fluxes.