Move call to wetdep_inputs_unset in bulk_aero/aero_model.F90
The call to the new routine wetdep_inputs_unset was
accidentally placed in the wrong subroutine (aero_model_surfarea
instead of aero_model_wetdep) in bulk_aero/aero_model.F90.
(It had been added correctly to modal_aero/aero_model.F90, which was
used in the initial testing). This error has been corrected,
and the update then tested, verifying that it now works.

[BFB]