SCREAM: improved atm dag verbosity options
Switched label from string to html, to increase flexibility.
The dag now prints with 5 levels of verbosity:

1: just atm procs names
2: atm procs names; fields names
3: atm procs names; fields names, tags, and dims
4: atm procs names; fields names, tags, dims, and grid name
5: atm procs names: fields names, tags, dims, units, and grid name

The driver assumes verbosity 5 in case of missing dependencies in the
dag (error), unless differently specified in the Debug sublist.
For writing dag (even if no error encountered), specify a verbosity
of at least 1 in the Debug sublist.