Replace psetcols with pcols
Globally replace psetcols with pcols, removing psetcols
as a member of the type physics_state, physics_tend,
and physics_ptend. Also remove psetcols/pcols as
a parameter to a number of routines, hardcoding the
value to pcols inside of the routines. This allows
a performance evaluation of restoring a more
global use of compile-time dimensioning of the first
dimension in chunks.

This breaks the logic associated with subcolumns.
Where necessary to allow non-subcol runs to work,
this logic has been commented out in the code.

psetcols (and subcols) is still mentioned in the
comments in a number of places, so this does not
represent an attempt to eliminate all subcolumn
support for E3SM (it only breaks it).

[BFB]